Исправил ошибки в миграции БД.

+++ src/main/resources/db/migration/V1__create_tables.sql
@@ -20,7 +20,7 @@ CREATE TABLE order_item
     id                BIGSERIAL NOT NULL PRIMARY KEY,
     coffee_variety_id BIGSERIAL NOT NULL,
     cups              INTEGER   NOT NULL,
-    order_id          integer   NOT NULL,
+    order_id          BIGSERIAL NOT NULL,
     FOREIGN KEY (coffee_variety_id) REFERENCES coffee_variety (id),
     FOREIGN KEY (order_id) REFERENCES orders (id)
 );
